+++ src/components/Carousel/index.tsx
@@ -6,7 +6,7 @@ import styles from './style.module.css';
 
 import { LeftArrow } from '~/components/ui/Svg/LeftArrow';
 import { RightArrow } from '~/components/ui/Svg/RightArrow';
-import { classNames } from '~/utils/classNames';
+import { classNames } from '~/libs/classNames';
 
 export type CarouselProps = {
 	components: ReactNode[];
