--- src/components/ui/Del/style.module.css
@@ -1,3 +1,0 @@
-.del {
-  text-decoration: line-through;
-}
